requirements.txt
fastapi==0.115.0
uvicorn==0.30.6
pydantic==2.9.2
SQLAlchemy==2.0.35
psycopg[binary]==3.2.1
pandas==2.2.2
python-dateutil==2.9.0.post0
pytest==8.3.3
httpx==0.27.2
python-dotenv==1.0.1